Fixed casting warnings on linux and macOS

+++ resolver.c
@@ -92,7 +92,7 @@ static bool proxy_resolver_get_proxies_for_url_from_system_config(void *ctx, con
         } else {
             // Construct proxy list url using scheme associated with proxy's port if available,
             // otherwise continue to use scheme associated with the url.
-            const int32_t proxy_port = get_host_port(proxy, strlen(proxy), 0);
+            const uint16_t proxy_port = get_host_port(proxy, strlen(proxy), 0);
             const char *proxy_scheme = proxy_port ? get_port_scheme(proxy_port, scheme) : scheme;
 
             // Use proxy from settings
